Move API to DebuggingSession
