+++ src/Features/Core/Portable/EditAndContinue/DebuggingSession.cs
@@ -18,6 +18,7 @@ using Microsoft.CodeAnalysis.ErrorReporting;
 using Microsoft.CodeAnalysis.NavigateTo;
 using Microsoft.CodeAnalysis.PooledObjects;
 using Microsoft.CodeAnalysis.Shared.Extensions;
+using Microsoft.CodeAnalysis.Text;
 using Microsoft.VisualStudio.Debugger.Contracts.EditAndContinue;
 using Roslyn.Utilities;
 
@@ -480,6 +481,434 @@ namespace Microsoft.CodeAnalysis.EditAndContinue
             {
                 return ImmutableArray<Diagnostic>.Empty;
             }
+        }
+
+        public async ValueTask<EmitSolutionUpdateResults> EmitSolutionUpdateAsync(
+            Solution solution,
+            ActiveStatementSpanProvider activeStatementSpanProvider,
+            CancellationToken cancellationToken)
+        {
+            var solutionUpdate = await EditSession.EmitSolutionUpdateAsync(solution, activeStatementSpanProvider, cancellationToken).ConfigureAwait(false);
+            if (solutionUpdate.ModuleUpdates.Status == ManagedModuleUpdateStatus.Ready)
+            {
+                StorePendingUpdate(solution, solutionUpdate);
+            }
+
+            // Note that we may return empty deltas if all updates have been deferred.
+            // The debugger will still call commit or discard on the update batch.
+            return new EmitSolutionUpdateResults(solutionUpdate.ModuleUpdates, solutionUpdate.Diagnostics, solutionUpdate.DocumentsWithRudeEdits);
+        }
+
+        public void CommitSolutionUpdate(out ImmutableArray<DocumentId> documentsToReanalyze)
+        {
+            var pendingUpdate = RetrievePendingUpdate();
+            CommitSolutionUpdate(pendingUpdate);
+
+            // restart edit session with no active statements (switching to run mode):
+            RestartEditSession(inBreakState: false, out documentsToReanalyze);
+        }
+
+        public void DiscardSolutionUpdate()
+            => _ = RetrievePendingUpdate();
+
+        public async ValueTask<ImmutableArray<ImmutableArray<ActiveStatementSpan>>> GetBaseActiveStatementSpansAsync(Solution solution, ImmutableArray<DocumentId> documentIds, CancellationToken cancellationToken)
+        {
+            if (!EditSession.InBreakState)
+            {
+                return default;
+            }
+
+            var baseActiveStatements = await EditSession.BaseActiveStatements.GetValueAsync(cancellationToken).ConfigureAwait(false);
+            using var _1 = PooledDictionary<string, ArrayBuilder<(ProjectId, int)>>.GetInstance(out var documentIndicesByMappedPath);
+            using var _2 = PooledHashSet<ProjectId>.GetInstance(out var projectIds);
+
+            // Construct map of mapped file path to a text document in the current solution
+            // and a set of projects these documents are contained in.
+            for (var i = 0; i < documentIds.Length; i++)
+            {
+                var documentId = documentIds[i];
+
+                var document = await solution.GetTextDocumentAsync(documentId, cancellationToken).ConfigureAwait(false);
+                if (document?.FilePath == null)
+                {
+                    // document has been deleted or has no path (can't have an active statement anymore):
+                    continue;
+                }
+
+                // Multiple documents may have the same path (linked file).
+                // The documents represent the files that #line directives map to.
+                // Documents that have the same path must have different project id.
+                documentIndicesByMappedPath.MultiAdd(document.FilePath, (documentId.ProjectId, i));
+                projectIds.Add(documentId.ProjectId);
+            }
+
+            using var _3 = PooledDictionary<ActiveStatement, ArrayBuilder<(DocumentId unmappedDocumentId, LinePositionSpan span)>>.GetInstance(
+                out var activeStatementsInChangedDocuments);
+
+            // Analyze changed documents in projects containing active statements:
+            foreach (var projectId in projectIds)
+            {
+                var newProject = solution.GetRequiredProject(projectId);
+                var analyzer = newProject.LanguageServices.GetRequiredService<IEditAndContinueAnalyzer>();
+
+                await foreach (var documentId in EditSession.GetChangedDocumentsAsync(LastCommittedSolution, newProject, cancellationToken).ConfigureAwait(false))
+                {
+                    cancellationToken.ThrowIfCancellationRequested();
+
+                    var newDocument = await solution.GetRequiredDocumentAsync(documentId, includeSourceGenerated: true, cancellationToken).ConfigureAwait(false);
+
+                    var (oldDocument, _) = await LastCommittedSolution.GetDocumentAndStateAsync(newDocument.Id, newDocument, cancellationToken).ConfigureAwait(false);
+                    if (oldDocument == null)
+                    {
+                        // Document is out-of-sync, can't reason about its content with respect to the binaries loaded in the debuggee.
+                        continue;
+                    }
+
+                    var oldDocumentActiveStatements = await baseActiveStatements.GetOldActiveStatementsAsync(analyzer, oldDocument, cancellationToken).ConfigureAwait(false);
+
+                    var analysis = await analyzer.AnalyzeDocumentAsync(
+                        LastCommittedSolution.GetRequiredProject(documentId.ProjectId),
+                        baseActiveStatements,
+                        newDocument,
+                        newActiveStatementSpans: ImmutableArray<LinePositionSpan>.Empty,
+                        Capabilities,
+                        cancellationToken).ConfigureAwait(false);
+
+                    // Document content did not change or unable to determine active statement spans in a document with syntax errors:
+                    if (!analysis.ActiveStatements.IsDefault)
+                    {
+                        for (var i = 0; i < oldDocumentActiveStatements.Length; i++)
+                        {
+                            // Note: It is possible that one active statement appears in multiple documents if the documents represent a linked file.
+                            // Example (old and new contents):
+                            //   #if Condition       #if Condition
+                            //     #line 1 a.txt       #line 1 a.txt
+                            //     [|F(1);|]           [|F(1000);|]     
+                            //   #else               #else
+                            //     #line 1 a.txt       #line 1 a.txt
+                            //     [|F(2);|]           [|F(2);|]
+                            //   #endif              #endif
+                            // 
+                            // In the new solution the AS spans are different depending on which document view of the same file we are looking at.
+                            // Different views correspond to different projects.
+                            activeStatementsInChangedDocuments.MultiAdd(oldDocumentActiveStatements[i].Statement, (analysis.DocumentId, analysis.ActiveStatements[i].Span));
+                        }
+                    }
+                }
+            }
+
+            using var _4 = ArrayBuilder<ImmutableArray<ActiveStatementSpan>>.GetInstance(out var spans);
+            spans.AddMany(ImmutableArray<ActiveStatementSpan>.Empty, documentIds.Length);
+
+            foreach (var (mappedPath, documentBaseActiveStatements) in baseActiveStatements.DocumentPathMap)
+            {
+                if (documentIndicesByMappedPath.TryGetValue(mappedPath, out var indices))
+                {
+                    // translate active statements from base solution to the new solution, if the documents they are contained in changed:
+                    foreach (var (projectId, index) in indices)
+                    {
+                        spans[index] = documentBaseActiveStatements.SelectAsArray(
+                            activeStatement =>
+                            {
+                                LinePositionSpan span;
+                                DocumentId? unmappedDocumentId;
+
+                                if (activeStatementsInChangedDocuments.TryGetValue(activeStatement, out var newSpans))
+                                {
+                                    (unmappedDocumentId, span) = newSpans.Single(ns => ns.unmappedDocumentId.ProjectId == projectId);
+                                }
+                                else
+                                {
+                                    span = activeStatement.Span;
+                                    unmappedDocumentId = null;
+                                }
+
+                                return new ActiveStatementSpan(activeStatement.Ordinal, span, activeStatement.Flags, unmappedDocumentId);
+                            });
+                    }
+                }
+            }
+
+            documentIndicesByMappedPath.FreeValues();
+            activeStatementsInChangedDocuments.FreeValues();
+
+            return spans.ToImmutable();
+        }
+
+        public async ValueTask<ImmutableArray<ActiveStatementSpan>> GetAdjustedActiveStatementSpansAsync(TextDocument mappedDocument, ActiveStatementSpanProvider activeStatementSpanProvider, CancellationToken cancellationToken)
+        {
+            if (!EditSession.InBreakState)
+            {
+                return ImmutableArray<ActiveStatementSpan>.Empty;
+            }
+
+            if (!mappedDocument.State.SupportsEditAndContinue())
+            {
+                return ImmutableArray<ActiveStatementSpan>.Empty;
+            }
+
+            Contract.ThrowIfNull(mappedDocument.FilePath);
+
+            var newProject = mappedDocument.Project;
+            var newSolution = newProject.Solution;
+            var oldProject = LastCommittedSolution.GetProject(newProject.Id);
+            if (oldProject == null)
+            {
+                // project has been added, no changes in active statement spans:
+                return ImmutableArray<ActiveStatementSpan>.Empty;
+            }
+
+            var baseActiveStatements = await EditSession.BaseActiveStatements.GetValueAsync(cancellationToken).ConfigureAwait(false);
+            if (!baseActiveStatements.DocumentPathMap.TryGetValue(mappedDocument.FilePath, out var oldMappedDocumentActiveStatements))
+            {
+                // no active statements in this document
+                return ImmutableArray<ActiveStatementSpan>.Empty;
+            }
+
+            var newDocumentActiveStatementSpans = await activeStatementSpanProvider(mappedDocument.Id, mappedDocument.FilePath, cancellationToken).ConfigureAwait(false);
+            if (newDocumentActiveStatementSpans.IsEmpty)
+            {
+                return ImmutableArray<ActiveStatementSpan>.Empty;
+            }
+
+            var analyzer = newProject.LanguageServices.GetRequiredService<IEditAndContinueAnalyzer>();
+
+            using var _ = ArrayBuilder<ActiveStatementSpan>.GetInstance(out var adjustedMappedSpans);
+
+            // Start with the current locations of the tracking spans.
+            adjustedMappedSpans.AddRange(newDocumentActiveStatementSpans);
+
+            // Update tracking spans to the latest known locations of the active statements contained in changed documents based on their analysis.
+            await foreach (var unmappedDocumentId in EditSession.GetChangedDocumentsAsync(LastCommittedSolution, newProject, cancellationToken).ConfigureAwait(false))
+            {
+                var newUnmappedDocument = await newSolution.GetRequiredDocumentAsync(unmappedDocumentId, includeSourceGenerated: true, cancellationToken).ConfigureAwait(false);
+
+                var (oldUnmappedDocument, _) = await LastCommittedSolution.GetDocumentAndStateAsync(newUnmappedDocument.Id, newUnmappedDocument, cancellationToken).ConfigureAwait(false);
+                if (oldUnmappedDocument == null)
+                {
+                    // document out-of-date
+                    continue;
+                }
+
+                var analysis = await EditSession.Analyses.GetDocumentAnalysisAsync(LastCommittedSolution, oldUnmappedDocument, newUnmappedDocument, activeStatementSpanProvider, Capabilities, cancellationToken).ConfigureAwait(false);
+
+                // Document content did not change or unable to determine active statement spans in a document with syntax errors:
+                if (!analysis.ActiveStatements.IsDefault)
+                {
+                    foreach (var activeStatement in analysis.ActiveStatements)
+                    {
+                        var i = adjustedMappedSpans.FindIndex((s, ordinal) => s.Ordinal == ordinal, activeStatement.Ordinal);
+                        if (i >= 0)
+                        {
+                            adjustedMappedSpans[i] = new ActiveStatementSpan(activeStatement.Ordinal, activeStatement.Span, activeStatement.Flags, unmappedDocumentId);
+                        }
+                    }
+                }
+            }
+
+            return adjustedMappedSpans.ToImmutable();
+        }
+
+        public async ValueTask<LinePositionSpan?> GetCurrentActiveStatementPositionAsync(Solution solution, ActiveStatementSpanProvider activeStatementSpanProvider, ManagedInstructionId instructionId, CancellationToken cancellationToken)
+        {
+            try
+            {
+                // It is allowed to call this method before entering or after exiting break mode. In fact, the VS debugger does so.
+                // We return null since there the concept of active statement only makes sense during break mode.
+                if (!EditSession.InBreakState)
+                {
+                    return null;
+                }
+
+                var baseActiveStatements = await EditSession.BaseActiveStatements.GetValueAsync(cancellationToken).ConfigureAwait(false);
+                if (!baseActiveStatements.InstructionMap.TryGetValue(instructionId, out var baseActiveStatement))
+                {
+                    return null;
+                }
+
+                var documentId = await FindChangedDocumentContainingUnmappedActiveStatementAsync(baseActiveStatements, instructionId.Method.Module, baseActiveStatement, solution, cancellationToken).ConfigureAwait(false);
+                if (documentId == null)
+                {
+                    // Active statement not found in any changed documents, return its last position:
+                    return baseActiveStatement.Span;
+                }
+
+                var newDocument = await solution.GetDocumentAsync(documentId, includeSourceGenerated: true, cancellationToken).ConfigureAwait(false);
+                if (newDocument == null)
+                {
+                    // The document has been deleted.
+                    return null;
+                }
+
+                var (oldDocument, _) = await LastCommittedSolution.GetDocumentAndStateAsync(newDocument.Id, newDocument, cancellationToken).ConfigureAwait(false);
+                if (oldDocument == null)
+                {
+                    // document out-of-date
+                    return null;
+                }
+
+                var analysis = await EditSession.Analyses.GetDocumentAnalysisAsync(LastCommittedSolution, oldDocument, newDocument, activeStatementSpanProvider, Capabilities, cancellationToken).ConfigureAwait(false);
+                if (!analysis.HasChanges)
+                {
+                    // Document content did not change:
+                    return baseActiveStatement.Span;
+                }
+
+                if (analysis.HasSyntaxErrors)
+                {
+                    // Unable to determine active statement spans in a document with syntax errors:
+                    return null;
+                }
+
+                Contract.ThrowIfTrue(analysis.ActiveStatements.IsDefault);
+                return analysis.ActiveStatements.GetStatement(baseActiveStatement.Ordinal).Span;
+            }
+            catch (Exception e) when (FatalError.ReportAndCatchUnlessCanceled(e, cancellationToken))
+            {
+                return null;
+            }
+        }
+
+        /// <summary>
+        /// Called by the debugger to determine whether a non-leaf active statement is in an exception region,
+        /// so it can determine whether the active statement can be remapped. This only happens when the EnC is about to apply changes.
+        /// If the debugger determines we can remap active statements, the application of changes proceeds.
+        /// 
+        /// TODO: remove (https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1310859)
+        /// </summary>
+        /// <returns>
+        /// True if the instruction is located within an exception region, false if it is not, null if the instruction isn't an active statement in a changed method 
+        /// or the exception regions can't be determined.
+        /// </returns>
+        public async ValueTask<bool?> IsActiveStatementInExceptionRegionAsync(Solution solution, ManagedInstructionId instructionId, CancellationToken cancellationToken)
+        {
+            try
+            {
+                if (!EditSession.InBreakState)
+                {
+                    return null;
+                }
+
+                // This method is only called when the EnC is about to apply changes, at which point all active statements and
+                // their exception regions will be needed. Hence it's not necessary to scope this query down to just the instruction
+                // the debugger is interested at this point while not calculating the others.
+
+                var baseActiveStatements = await EditSession.BaseActiveStatements.GetValueAsync(cancellationToken).ConfigureAwait(false);
+                if (!baseActiveStatements.InstructionMap.TryGetValue(instructionId, out var baseActiveStatement))
+                {
+                    return null;
+                }
+
+                var documentId = await FindChangedDocumentContainingUnmappedActiveStatementAsync(baseActiveStatements, instructionId.Method.Module, baseActiveStatement, solution, cancellationToken).ConfigureAwait(false);
+                if (documentId == null)
+                {
+                    // the active statement is contained in an unchanged document, thus it doesn't matter whether it's in an exception region or not
+                    return null;
+                }
+
+                var newDocument = solution.GetRequiredDocument(documentId);
+                var (oldDocument, _) = await LastCommittedSolution.GetDocumentAndStateAsync(newDocument.Id, newDocument, cancellationToken).ConfigureAwait(false);
+                if (oldDocument == null)
+                {
+                    // Document is out-of-sync, can't reason about its content with respect to the binaries loaded in the debuggee.
+                    return null;
+                }
+
+                var analyzer = newDocument.Project.LanguageServices.GetRequiredService<IEditAndContinueAnalyzer>();
+                var oldDocumentActiveStatements = await baseActiveStatements.GetOldActiveStatementsAsync(analyzer, oldDocument, cancellationToken).ConfigureAwait(false);
+                return oldDocumentActiveStatements.GetStatement(baseActiveStatement.Ordinal).ExceptionRegions.IsActiveStatementCovered;
+            }
+            catch (Exception e) when (FatalError.ReportAndCatchUnlessCanceled(e, cancellationToken))
+            {
+                return null;
+            }
+        }
+
+        private async Task<DocumentId?> FindChangedDocumentContainingUnmappedActiveStatementAsync(
+            ActiveStatementsMap activeStatementsMap,
+            Guid moduleId,
+            ActiveStatement baseActiveStatement,
+            Solution newSolution,
+            CancellationToken cancellationToken)
+        {
+            DocumentId? documentId = null;
+            if (TryGetProjectId(moduleId, out var projectId))
+            {
+                var oldProject = LastCommittedSolution.GetProject(projectId);
+                if (oldProject == null)
+                {
+                    // project has been added (should have no active statements under normal circumstances)
+                    return null;
+                }
+
+                var newProject = newSolution.GetProject(projectId);
+                if (newProject == null)
+                {
+                    // project has been deleted
+                    return null;
+                }
+
+                documentId = await GetChangedDocumentContainingUnmappedActiveStatementAsync(activeStatementsMap, LastCommittedSolution, newProject, baseActiveStatement, cancellationToken).ConfigureAwait(false);
+            }
+            else
+            {
+                // Search for the document in all changed projects in the solution.
+
+                using var documentFoundCancellationSource = new CancellationTokenSource();
+                using var linkedTokenSource = CancellationTokenSource.CreateLinkedTokenSource(documentFoundCancellationSource.Token, cancellationToken);
+
+                async Task GetTaskAsync(ProjectId projectId)
+                {
+                    var newProject = newSolution.GetRequiredProject(projectId);
+                    var id = await GetChangedDocumentContainingUnmappedActiveStatementAsync(activeStatementsMap, LastCommittedSolution, newProject, baseActiveStatement, linkedTokenSource.Token).ConfigureAwait(false);
+                    Interlocked.CompareExchange(ref documentId, id, null);
+                    if (id != null)
+                    {
+                        documentFoundCancellationSource.Cancel();
+                    }
+                }
+
+                var tasks = newSolution.ProjectIds.Select(GetTaskAsync);
+
+                try
+                {
+                    await Task.WhenAll(tasks).ConfigureAwait(false);
+                }
+                catch (OperationCanceledException) when (documentFoundCancellationSource.IsCancellationRequested)
+                {
+                    // nop: cancelled because we found the document
+                }
+            }
+
+            return documentId;
+        }
+
+        // Enumerate all changed documents in the project whose module contains the active statement.
+        // For each such document enumerate all #line directives to find which maps code to the span that contains the active statement.
+        private static async ValueTask<DocumentId?> GetChangedDocumentContainingUnmappedActiveStatementAsync(ActiveStatementsMap baseActiveStatements, CommittedSolution oldSolution, Project newProject, ActiveStatement activeStatement, CancellationToken cancellationToken)
+        {
+            var analyzer = newProject.LanguageServices.GetRequiredService<IEditAndContinueAnalyzer>();
+
+            await foreach (var documentId in EditSession.GetChangedDocumentsAsync(oldSolution, newProject, cancellationToken).ConfigureAwait(false))
+            {
+                cancellationToken.ThrowIfCancellationRequested();
+
+                var newDocument = newProject.GetRequiredDocument(documentId);
+                var (oldDocument, _) = await oldSolution.GetDocumentAndStateAsync(newDocument.Id, newDocument, cancellationToken).ConfigureAwait(false);
+                if (oldDocument == null)
+                {
+                    // Document is out-of-sync, can't reason about its content with respect to the binaries loaded in the debuggee.
+                    return null;
+                }
+
+                var oldActiveStatements = await baseActiveStatements.GetOldActiveStatementsAsync(analyzer, oldDocument, cancellationToken).ConfigureAwait(false);
+                if (oldActiveStatements.Any(s => s.Statement == activeStatement))
+                {
+                    return documentId;
+                }
+            }
+
+            return null;
         }
 
         internal TestAccessor GetTestAccessor()
